applied SRP and Pure Fabrication to room and reservation classes. cleaned up imports.
