+++ spring-boot/src/main/java/springboot/Guest.java
@@ -1,7 +1,6 @@
 package springboot;
 
 import springboot.database.Setup;
-
 import java.sql.*;
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -50,29 +49,6 @@ public class Guest implements User {
 
     public void setStreetAddress(String streetAddress) {
         this.streetAddress = streetAddress;
-    }
-
-    // User class interface methods
-
-    //returns either a failure message or "success"
-    public String reserveRoom(Room reservedRoom, LocalDate start, LocalDate end) {
-        if(reservedRoom != null) {
-            Reservation res = new Reservation(reservedRoom, start, end);
-
-            return Reservation.addToDatabase(res.getStartDay(),
-                    res.getEndDay(),
-                    res.getRoom().getRoomNumber(),
-                    res.getUsername());
-        }
-        else{
-            return "room does not exist";
-        }
-    }
-
-    public String reserveRoom(Room reservedRoom, String start, String end) {
-        return reserveRoom(reservedRoom,
-                DateParsing.convertStringToDate(start),
-                DateParsing.convertStringToDate(end));
     }
