+++ spring-boot/src/main/java/springboot/controller/AccountController.java
@@ -1,7 +1,5 @@
 package springboot.controller;
 
-import org.apache.juli.logging.Log;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import springboot.LoggedIn;
