+++ spring-boot/src/main/java/springboot/controller/AddRoomController.java
@@ -1,6 +1,5 @@
 package springboot.controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import springboot.service.RoomService;
