+++ spring-boot/src/main/java/springboot/controller/ProductController.java
@@ -3,7 +3,6 @@
 package springboot.controller;
 
 //Changed Imports
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import springboot.Product;
