+++ spring-boot/src/main/java/springboot/service/ProductService.java
@@ -2,7 +2,6 @@
 
 package springboot.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import springboot.LoggedIn;
 import springboot.Product;
